event tracks controller - implemented most event track management logic - added new routes to README.md - renamed /events/playlists route - refactoredEventTrack - Playlist entity relation - extended express request object with event type set in prepareEventAuthentication - fixed type check for changing participants permissions

+++ backend/src/index.ts
@@ -4,16 +4,11 @@ import {EntityManager, MikroORM, RequestContext} from '@mikro-orm/core';
 import {Auth} from "./middleware/auth.middleware";
 import {EventController} from "./controller/event.controller";
 import {SpotifyAuthController} from "./controller/auth.spotify.controller";
-import cors from "cors";
 import {SchemaGenerator} from "@mikro-orm/postgresql";
 
-const PORT = 4000;// TODO: move into env file
-const app = express();
+import 'dotenv/config';
 
-// TODO: move into env file
-const SPOTIFY_CLIENT_ID = "b241ae6416a3481dad98e6899b7be0b4";
-const SPOTIFY_CLIENT_SECRET = "4807da9be0f144b9bbab888217e5e969";
-const SPOTIFY_REDIRECT_URI = "http://localhost:4000/account/login_response";
+const app = express();
 
 export const DI = {} as {
     server: http.Server;
@@ -27,28 +22,34 @@ export const DI = {} as {
 };
 
 export const initializeServer = async () => {
+    if (process.env.SPOTIFY_CLIENT_ID == undefined
+        || process.env.SPOTIFY_CLIENT_SECRET == undefined
+        || process.env.SPOTIFY_REDIRECT_URI == undefined) {
+        console.log("Error: Make sure SPOTIFY_CLIENT_ID, SPOTIFY_CLIENT_SECRET and SPOTIFY_REDIRECT_URI is set in env.");
+        process.exit(1);
+    }
+
     // dependency injection setup
     DI.orm = await MikroORM.init();
     DI.em = DI.orm.em;
     DI.generator = await <SchemaGenerator>DI.orm.getSchemaGenerator();
-    DI.spotifyClientId = SPOTIFY_CLIENT_ID;
-    DI.spotifyClientSecret = SPOTIFY_CLIENT_SECRET;
-    DI.spotifyRedirectUri = SPOTIFY_REDIRECT_URI;
+    DI.spotifyClientId = process.env.SPOTIFY_CLIENT_ID;
+    DI.spotifyClientSecret = process.env.SPOTIFY_CLIENT_SECRET;
+    DI.spotifyRedirectUri = process.env.SPOTIFY_REDIRECT_URI;
     DI.frontendUrl = "http://localhost:5173";
 
-    if (process.env.environment == 'test') {
-        await DI.generator.dropSchema();
-    }
-
+    if (process.env.environment == 'test') await DI.generator.dropSchema();
     await DI.generator.updateSchema();
-    app.all('/*', function (req, res, next) {
+    console.log("All database schemas updated !");
+
+    /*app.all('/*', function (req, res, next) {
         res.header("Access-Control-Allow-Origin", "*");
         res.header("Access-Control-Allow-Headers", "X-Requested-With");
         next();
     });
     app.use(cors({
         origin: '*', preflightContinue: true, optionsSuccessStatus: 200
-    }));
+    }));*/
 
     // example middleware
     app.use((req, res, next) => {
@@ -67,12 +68,11 @@ export const initializeServer = async () => {
     app.use('/account', SpotifyAuthController);
     app.use('/events', Auth.verifySpotifyAccess, EventController);
 
-    DI.server = app.listen(PORT, () => {
-        console.log(`Server started on port ${PORT}`);
+    DI.server = app.listen(process.env.PORT, () => {
+        console.log(`Server started on port ${process.env.PORT}`);
     });
 };
 
-// TODO: create env file
 if (process.env.environment !== 'test') {
     initializeServer();
 }
